--- src/app/views/settings/index.ts
@@ -1,3 +1,0 @@
-import { Settings } from './Settings';
-
-export { Settings };
